separated concerns and added world class
