+++ README.md
@@ -1,2 +1,2 @@
 # The_Haunting_Of_Amazon_Hill
-Text Based Adventure Map
+Text Based Adventure World
