+++ src/Main.java
@@ -1,6 +1,4 @@
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 
 public class Main {
     public static void main(String[] args) throws IOException {
@@ -13,3 +11,4 @@ public class Main {
         s.splash();
     }
 }
+
